Hash join node codegen.

+++ be/src/util/hash-util.h
@@ -24,7 +24,7 @@ class HashUtil {
   // the current hash/seed value.
   // This should only be called if SSE is supported.
   // This is ~4x faster than Fvn/Boost Hash.
-  static uint32_t CrcHash(const void* data, size_t bytes, uint32_t hash) {
+  static uint32_t CrcHash(const void* data, int32_t bytes, uint32_t hash) {
     DCHECK(CpuInfo::Instance()->IsSupported(CpuInfo::SSE4_2));
     uint32_t words = bytes / sizeof(uint32_t);
     bytes = bytes % sizeof(uint32_t);
@@ -55,7 +55,7 @@ class HashUtil {
   // For example, if the data is <1000, 2000, 3000, 4000, ..> and then the mod of 1000 
   // is taken on the hash, all values will collide to the same bucket.
   // For string values, Fvn is slightly faster than boost.
-  static uint32_t FvnHash(const void* data, size_t bytes, uint32_t hash) {
+  static uint32_t FvnHash(const void* data, int32_t bytes, uint32_t hash) {
     const uint8_t* ptr = reinterpret_cast<const uint8_t*>(data);
     while (bytes--) {
       hash = (*ptr ^ hash) * FVN_PRIME;
@@ -66,7 +66,7 @@ class HashUtil {
 
   // Computes the hash value for data.  Will call either CrcHash or FvnHash
   // depending on hardware capabilities.
-  static uint32_t Hash(const void* data, size_t bytes, uint32_t hash) {
+  static uint32_t Hash(const void* data, int32_t bytes, uint32_t hash) {
 #ifdef __SSE4_2__
     if (CpuInfo::Instance()->IsSupported(CpuInfo::SSE4_2)) {
       return CrcHash(data, bytes, hash);
